🗂️ chore: organize project structure by moving custom components to customComponents folder
