+++ src/app/(auth)/signup/SignUpForm.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+import LoadingButton from "@/components/customComponents/LoadingButton";
+import { PasswordInput } from "@/components/customComponents/PasswordInput";
 import {
   Form,
   FormControl,
@@ -9,8 +11,6 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import LoadingButton from "@/components/ui/LoadingButton";
-import { PasswordInput } from "@/components/ui/PasswordInput";
 import { validationsMessages } from "@/lib/constants";
 import { getSignUpSchema, SignUpValuesType } from "@/lib/validations";
 import { zodResolver } from "@hookform/resolvers/zod";
