+++ src/components/customComponents/SearchField.tsx
@@ -3,7 +3,7 @@
 import { SearchIcon } from "lucide-react";
 import { useRouter } from "next/navigation";
 import React from "react";
-import { Input } from "./ui/input";
+import { Input } from "../ui/input";
 
 export default function SearchField() {
   const router = useRouter();
